Add auth functionality
This commit adds buttons on the navbar for login, register, and log out.
This commit also adds the properties isLoggedIn and isAdmin to all views
rendered by the application to suppor the navbar. This commit also adds
auth guards to the admin page and redirects users if they try to access
an auth screen once already logged in.